pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: BSD License",
]
name = "nomad-laserphysics"
description = "Nomad example template"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Kai Koebnick", email = "kai.koebnick@fau.de" },
]
maintainers = [
    { name = "Kai Koebnick", email = "kai.koebnick@fau.de" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.3.0",
    "python-magic-bin; sys_platform == 'win32'",
]

[project.urls]
Repository = "https://github.com/kaikoebnick/nomad-laserphysics"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
measurements_app_entry_point = "nomad_laserphysics.apps:measurements_app_entry_point"
FEM_correlation_chambers_app_entry_point = "nomad_laserphysics.apps:FEM_correlation_chambers_app_entry_point"
FIM_test_chambers_app_entry_point = "nomad_laserphysics.apps:FIM_test_chambers_app_entry_point"
evaluations_app_entry_point = "nomad_laserphysics.apps:evaluations_app_entry_point"
objects_app_entry_point = "nomad_laserphysics.apps:objects_app_entry_point"
tip_samples_app_entry_point = "nomad_laserphysics.apps:tip_samples_app_entry_point"
evaluation_schema_package_entry_point = "nomad_laserphysics.schema_packages:evaluation_schema_package_entry_point"
ML_evaluation_schema_package_entry_point = "nomad_laserphysics.schema_packages:ML_evaluation_schema_package_entry_point"
ML_evaluation_procedure_schema_package_entry_point = "nomad_laserphysics.schema_packages:ML_evaluation_procedure_schema_package_entry_point"
measurement_schema_package_entry_point = "nomad_laserphysics.schema_packages:measurement_schema_package_entry_point"
FIM_test_chamber_schema_package_entry_point = "nomad_laserphysics.schema_packages:FIM_test_chamber_schema_package_entry_point"
FEM_correlation_chamber_schema_package_entry_point = "nomad_laserphysics.schema_packages:FEM_correlation_chamber_schema_package_entry_point"
object_schema_package_entry_point = "nomad_laserphysics.schema_packages:object_schema_package_entry_point"
tip_sample_schema_package_entry_point = "nomad_laserphysics.schema_packages:tip_sample_schema_package_entry_point"

[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
